requirements.txt
django
djangorestframework
pyyaml
requests
django-cors-headers
pillow
django-mptt